pyproject.toml
[tool.poetry]
name = "shoppets"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
psycopg = {extras = ["binary"], version = "^3.1.17"}
django-mptt = "^0.16.0"
django-mptt-admin = "^2.6.1"
pytils = "^0.4.1"
django-debug-toolbar = "^4.3.0"
djoser = "^2.2.2"
djangorestframework-simplejwt = "^5.3.1"
social-auth-app-django = "^5.4.0"
drf-spectacular = "^0.27.1"
black = "^24.2.0"
ipython = "^8.22.2"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
